data-ml-pipelines/projects/item_ranking/content_based/requirements.txt
cloudpickle==3.1.0
Jinja2==3.1.5
joblib==1.4.2
mlflow==2.16.0
mlflow_skinny==2.16.0
numpy==1.26.0
pandas==2.2.3
protobuf==5.29.2
python-dotenv==1.0.1
sentence_transformers==3.3.1
sentence-transformers[onnx]
structlog==24.4.0
torch==2.2.2
google-cloud-bigquery==3.26.0
google-cloud-storage==2.18.2
google-cloud-bigquery-storage==2.27.0
db-dtypes==1.4.0